toh: final solutions are now runnable for parts 1 and 2
