+++ public/docs/_examples/toh-1/ts-snippets/app.component.snippets.pt1.ts
