+++ public/docs/_examples/toh-1/ts/app/boot.ts
@@ -1,4 +1,6 @@
-import {bootstrap}    from 'angular2/platform/browser'
-import {AppComponent} from './app.component'
+// #docregion pt1
+import {bootstrap} from 'angular2/platform/browser';
+import {AppComponent} from './app.component';
 
 bootstrap(AppComponent);
+// #enddocregion pt1
