+++ public/docs/_examples/toh-2/ts-snippets/app.component.snippets.pt2.ts
