--- public/docs/_examples/toh-2/ts-snippets/app/boot.ts
@@ -1,6 +1,0 @@
-// #docregion pt1
-import {bootstrap} from 'angular2/platform/browser';
-import {AppComponent} from './app.component';
-
-bootstrap(AppComponent);
-// #enddocregion pt1
