+++ public/docs/_examples/toh-2/ts/app/app.component.ts
