+++ public/docs/_examples/toh-2/ts/plnkr.json
@@ -1,0 +1,8 @@
+{
+  "description": "Tour of Heroes: Part 2",
+  "files":[
+    "!**/*.d.ts",
+    "!**/*.js"
+  ],
+  "tags": ["tutorial", "tour", "heroes"]
+}
