+++ public/docs/ts/latest/tutorial/toh-pt1.jade
@@ -49,12 +49,12 @@ code-example(format="" language="bash").
     Let's add two properties to our `AppComponent`, a `title` property for the application name and a `hero` property
     for a hero named "Windstorm".
 
-  +makeExample('toh-1/ts-snippets/app/app.component.snippets.pt1.ts', 'app-component-1', 'app.component.ts (AppComponent class)')(format=".")
+  +makeExample('toh-1/ts-snippets/app.component.snippets.pt1.ts', 'app-component-1', 'app.component.ts (AppComponent class)')(format=".")
 
   :marked
     Now we update the template in the `@Component` decoration with data bindings to these new properties.
 
-  +makeExample('toh-1/ts-snippets/app/app.component.snippets.pt1.ts', 'show-hero')
+  +makeExample('toh-1/ts-snippets/app.component.snippets.pt1.ts', 'show-hero')
 
   :marked
     The browser should refresh and display our title and hero.
@@ -73,7 +73,7 @@ code-example(format="" language="bash").
   Create a `Hero` interface with `id` and `name` properties.
   Keep this near the top of the `app.component.ts` file for now.
 
-+makeExample('toh-1/ts-snippets/app/app.component.ts', 'hero-interface-1', 'app.component.ts (Hero interface)')(format=".")
++makeExample('toh-1/ts/app/app.component.ts', 'hero-interface-1', 'app.component.ts (Hero interface)')(format=".")
 
 .l-sub-section
   :marked
@@ -83,13 +83,13 @@ code-example(format="" language="bash").
   Now that we have a `Hero` interface, let’s refactor our component’s `hero` property to be of type `Hero`.
   Then initialize it with an id of `1` and the name, "Windstorm".
 
-+makeExample('toh-1/ts-snippets/app/app.component.snippets.pt1.ts', 'hero-property-1', 'app.component.ts (Hero property)')(format=".")
++makeExample('toh-1/ts-snippets/app.component.snippets.pt1.ts', 'hero-property-1', 'app.component.ts (Hero property)')(format=".")
 
 :marked
   Because we changed the hero from a string to an object,
   we update the binding in the template to refer to the hero’s `name` property.
 
-+makeExample('toh-1/ts-snippets/app/app.component.snippets.pt1.ts', 'show-hero-2')
++makeExample('toh-1/ts-snippets/app.component.snippets.pt1.ts', 'show-hero-2')
 :marked
   The browser refreshes and continues to display our hero’s name.
 
@@ -97,7 +97,7 @@ code-example(format="" language="bash").
   Displaying a name is good, but we want to see all of our hero’s properties.
   We’ll add a `<div>` for our hero’s `id` property and another `<div>` for our hero’s `name`.
 
-+makeExample('toh-1/ts-snippets/app/app.component.snippets.pt1.ts', 'show-hero-properties')
++makeExample('toh-1/ts-snippets/app.component.snippets.pt1.ts', 'show-hero-properties')
 :marked
   Uh oh, our template string is getting long. We better take care of that to avoid the risk of making a typo in the template.
 
@@ -112,7 +112,7 @@ code-example(format="" language="bash").
   Change the quotes around the template to back-ticks and
   put the `<h1>`, `<h2>` and `<div>` elements on their own lines.
 
-+makeExample('toh-1/ts-snippets/app/app.component.snippets.pt1.ts', 'multi-line-strings', 'app.component.ts (AppComponent\'s template)')
++makeExample('toh-1/ts-snippets/app.component.snippets.pt1.ts', 'multi-line-strings', 'app.component.ts (AppComponent\'s template)')
 
   .callout.is-important
     header A back-tick is not a single quote
@@ -132,7 +132,7 @@ code-example(format="" language="bash").
 
   Refactor the hero name `<label>` with `<label>` and `<input>` elements as shown below:
 
-+makeExample('toh-1/ts-snippets/app/app.component.snippets.pt1.ts', 'editing-Hero', 'app.component.ts (input element)')
++makeExample('toh-1/ts-snippets/app.component.snippets.pt1.ts', 'editing-Hero', 'app.component.ts (input element)')
 :marked
   We see in the browser that the hero’s name does appear in the `<input>` textbox.
   But something doesn’t feel right.
@@ -177,7 +177,7 @@ code-example(language="html").
 
   Here's the complete `app.component.ts` as it stands now:
 
-+makeExample('toh-1/ts-snippets/app/app.component.ts', 'pt1', 'app.component.ts')
++makeExample('toh-1/ts/app/app.component.ts', 'pt1', 'app.component.ts')
 
 .l-main-section
 :marked
